+++ src/argilla_server/apis/routes.py
@@ -42,6 +42,7 @@ from argilla_server.apis.v1.handlers import (
 from argilla_server.apis.v1.handlers import (
     fields as fields_v1,
 )
+from argilla_server.apis.v1.handlers import info as info_v1
 from argilla_server.apis.v1.handlers import (
     metadata_properties as metadata_properties_v1,
 )
@@ -114,6 +115,7 @@ def create_api_v1():
     APIErrorHandler.configure_app(api_v1)
 
     for router in [
+        info_v1.router,
         authentication_v1.router,
         datasets_v1.router,
         fields_v1.router,