+++ src/argilla_server/settings.py
@@ -128,9 +128,9 @@ class Settings(BaseSettings):
     )
 
     # Hugging Face settings
-    show_huggingface_space_persistant_storage_warning: bool = Field(
+    show_huggingface_space_persistent_storage_warning: bool = Field(
         default=True,
-        description="If True, show a warning when Hugging Face space persistant storage is disabled",
+        description="If True, show a warning when Hugging Face space persistent storage is disabled",
     )
 
     # See also the telemetry.py module