+++ CHANGELOG.md
@@ -17,6 +17,7 @@ These are the section headers that we use:
 ## [Unreleased]()
 
 - Added `POST /api/v1/token` endpoint to generate a new API token for a user. ([#138](https://github.com/argilla-io/argilla-server/pull/138))
+- Added `GET /api/v1/me` endpoint to get the current user information. ([#140](https://github.com/argilla-io/argilla-server/pull/140))
 
 ## [Unreleased]()